--- circuit-fault-diagnosis/CODEOWNERS
@@ -1,1 +1,0 @@
-* @bellert
