+++ circuit_fault_diagnosis/gates.py
