+++ tests/allcombos.py
